+++ NEWS.md
@@ -3,7 +3,9 @@
 ## Major Improvements
 
 * The left-hand side may now contain calls to `[[`, `[`, and `$` allowing
-  assignment of list elements and variables in specific environments.
+  assignment of list elements and variables in specific environments. The 
+  parent list, environment, or object in question must already exist, otherwise
+  an error is raised.
 
 # zeallot 0.0.5