+++ NEWS.md
@@ -3,9 +3,8 @@
 ## Major Improvements
 
 * The left-hand side may now contain calls to `[[`, `[`, and `$` allowing
-  assignment of list elements and variables in specific environments. The 
-  parent list, environment, or object in question must already exist, otherwise
-  an error is raised.
+assignment of parts of objects. The parent object in question must already
+exist, otherwise an error is raised. (@rafaqz, #32)
 
 # zeallot 0.0.5